fix(build): copy exact files, or best match when possible
Since we're generating the exact file as `prefix + name + ext` in the
build dir, we rely on that name to copy to output dir.

If that name is not present, because the user defined a user script like
make mostly (like in typescript), then take the first match and work
with that.

PR: https://github.com/stackmystack/tsdl/pull/22